Refactored game libraries
Each game class is now in its own source TypeScript file
