+++ src/direction.ts
@@ -1,0 +1,6 @@
+export enum Direction {
+	Down = 0,
+	Left,
+	Up,
+	Right	
+}
